Fix profile picture button
