+++ frontend/src/components/navbar/Navbar.jsx
@@ -6,19 +6,41 @@ import {
   faPeopleGroup,
   faCalendarWeek,
   faCirclePlus,
+  faRightFromBracket,
+  faGraduationCap,
 } from "@fortawesome/free-solid-svg-icons";
 import Image from "next/image";
 import LinkList from "@/components/navbar/LinkList";
 import Link from "next/link";
-import { useSession } from "next-auth/react";
+import { signOut, useSession } from "next-auth/react";
 import ProfilePicture from "@/components/ProfilePicture";
 import Logo from "/public/images/Logo-Dr-Trottoir-GEEL-01.png";
+import SecondaryButton from "@/components/button/SecondaryButton";
+import { useRouter } from "next/router";
+import { useEffect, useRef, useState } from "react";
 
 // https://flowbite.com/docs/components/sidebar/
 export default function Navbar() {
   const {
     data: { user },
   } = useSession();
+  const router = useRouter();
+  const [open, setOpen] = useState(false);
+  const ref = useRef(null);
+  const handleClickOutside = (event) => {
+    if (ref.current && !ref.current.contains(event.target)) {
+      setOpen(false);
+    }
+  };
+
+  useEffect(() => {
+    document.addEventListener("mousedown", handleClickOutside);
+
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, []);
+
   return (
     <div className={"flex flex-col flex-shrink-0 sm:w-64 h-full bg-dark-bg-1"}>
       <div className={"flex justify-center p-1 mb-2 mt-8"}>
@@ -63,7 +85,37 @@ export default function Navbar() {
           linkClassName={"hover: hover:bg-dark-bg-2"}
         />
       </div>
-      <div className={"flex p-8 pt-4 w-full items-center"} id={"p-info"}>
+      <div
+        className={"flex p-8 pt-4 w-full items-center relative cursor-pointer"}
+        id={"p-info"}
+        onClick={() => setOpen(true)}
+        ref={ref}
+      >
+        {open && (
+          <div
+            className={
+              "absolute bottom-[100%] absolute z-[100] hover:z-[1000] border-2 border-light-border mt-2 rounded-lg bg-light-bg-2 mb-4 w-[180px]"
+            }
+          >
+            <SecondaryButton
+              onClick={async () => {
+                await signOut({ redirect: false });
+                await router.push("/");
+              }}
+              icon={faRightFromBracket}
+              className={"w-full mb-1"}
+            >
+              Uitloggen
+            </SecondaryButton>
+            <SecondaryButton
+              onClick={() => router.push("/student/planning")}
+              icon={faGraduationCap}
+              className={"w-full"}
+            >
+              Naar Student
+            </SecondaryButton>
+          </div>
+        )}
         <ProfilePicture image={null} className={"w-8"} />
         <div className={"flex flex-col justify-center ml-6"}>
           <p className={"text-dark-text"}>{user.first_name}</p>
